+++ src/MonthValidator.php
@@ -30,7 +30,7 @@ final class MonthValidator extends FieldValidator
         '12' => 'DEC',
     ];
 
-    public function isSatisfiedBy(string $fieldExpression, DateTimeInterface $date): bool
+    protected function isSatisfiedExpression(string $fieldExpression, DateTimeInterface $date): bool
     {
         return '?' === $fieldExpression
             || $this->isSatisfied((int) $date->format('m'), $this->convertLiterals($fieldExpression));